Refactor the repository files
