+++ src/main/java/com/aegro/model/Production.java
@@ -6,6 +6,7 @@ import javax.persistence.Entity;
 import javax.persistence.Id;
 
 import org.springframework.data.mongodb.core.mapping.Document;
+import org.springframework.util.ObjectUtils;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
@@ -57,11 +58,8 @@ public class Production {
 	}
 	
 	@JsonIgnore
-	public boolean isNull() {
-		if(this.kilo == null && this.id ==null) {
-			return true;
-		}
-		return false;
+	public boolean isEmpty() {
+		return ObjectUtils.isEmpty(this);
 	}
 
 }
